Separate out EventGrid into its own module
