+++ azure-functions-sdk/src/commands/new.rs
@@ -16,11 +16,13 @@ use std::{
 use syn::{parse_file, Expr, ExprArray, Item};
 
 mod blob;
+mod event_grid;
 mod http;
 mod queue;
 mod timer;
 
 pub use self::blob::Blob;
+pub use self::event_grid::EventGrid;
 pub use self::http::Http;
 pub use self::queue::Queue;
 pub use self::timer::Timer;
@@ -269,41 +271,6 @@ impl<'a> From<&'a ArgMatches<'a>> for New<'a> {
             quiet: args.is_present("quiet"),
             color: args.value_of("color"),
             args,
-        }
-    }
-}
-
-struct EventGrid<'a> {
-    name: &'a str,
-}
-
-impl<'a> EventGrid<'a> {
-    pub fn create_subcommand<'b>() -> App<'a, 'b> {
-        SubCommand::with_name("event-grid")
-            .about("Creates a new Event Grid triggered Azure Function.")
-            .arg(
-                Arg::with_name("name")
-                    .long("name")
-                    .short("n")
-                    .value_name("NAME")
-                    .help("The name of the new Azure Function.")
-                    .required(true),
-            )
-    }
-
-    pub fn execute(&self, quiet: bool) -> Result<(), String> {
-        let data = json!({
-            "name": self.name,
-        });
-
-        create_function(self.name, "eventgrid.rs", &data, quiet)
-    }
-}
-
-impl<'a> From<&'a ArgMatches<'a>> for EventGrid<'a> {
-    fn from(args: &'a ArgMatches<'a>) -> Self {
-        EventGrid {
-            name: args.value_of("name").unwrap(),
         }
     }
 }
